requirements.txt
streamlit
tensorflow
torch
ultralytics
pillow
numpy
